pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riscv-parameter-extraction"
version = "0.1.0"
description = "A tool for extracting parameters from RISC-V specifications"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Balgopal Moharana", email = "lucifer4073@gmail.com"}
]
keywords = ["riscv", "parameter-extraction", "llm", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "litellm",
    "python-dotenv",
    "langchain",
    "langchain-community",
    "faiss-cpu",
    "sentence-transformers",
    "ollama",
    "pyyaml",
    "pydantic",
    "numpy",
]

[project.urls]
Repository = "https://github.com/lucifer4073/riscv_parameter_extraction"

[tool.setuptools]
packages = ["riscv_parameter_extraction"]

[tool.setuptools.package-data]
riscv_parameter_extraction = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
